requirements.txt
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
wikipedia
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
arxiv
wikipedia
streamlit
duckduckgo-search